[IMPR] Security in calender module

+++ src/Calender/Services/CalenderService.ts
@@ -1,6 +1,7 @@
-import { Pagination, Service } from '@core'
+import { Pagination, Service, UnauthorizedError } from '@core'
 import { CalenderEvent } from '../Data/CalenderEvent'
 import { CalenderEventRepository } from '../Data/CalenderEventRepository'
+import { User } from '@modules/Users/Data/User'
 
 export class CalenderService extends Service<CalenderEvent> {
 	private readonly calenderEventRepository: CalenderEventRepository
@@ -11,13 +12,21 @@ export class CalenderService extends Service<CalenderEvent> {
 		this.calenderEventRepository = new CalenderEventRepository()
 	}
 
-	public async create(event: CalenderEvent): Promise<void> {
-		await this.calenderEventRepository.create(event)
+	public async create(
+		event: CalenderEvent,
+		username: User['username']
+	): Promise<void> {
+		await this.calenderEventRepository.create({ ...event, username })
 	}
 
-	public async get(pagination?: Pagination): Promise<CalenderEvent[]> {
+	public async get(
+		username: User['username'],
+		pagination?: Pagination
+	): Promise<CalenderEvent[]> {
 		const events = this.calenderEventRepository.find(
-			undefined,
+			{
+				username,
+			},
 			undefined,
 			pagination
 		)
@@ -33,19 +42,46 @@ export class CalenderService extends Service<CalenderEvent> {
 	}
 
 	public async getOne(
-		id: CalenderEvent['id']
+		id: CalenderEvent['id'],
+		username: User['username']
 	): Promise<CalenderEvent | null> {
-		return this.calenderEventRepository.findOne({ id })
+		const event = await this.calenderEventRepository.findOne({ id })
+
+		if (event && event?.username !== username) {
+			throw new UnauthorizedError()
+		}
+
+		return event
 	}
 
 	public async update(
 		id: CalenderEvent['id'],
-		event: Partial<CalenderEvent>
+		event: Partial<CalenderEvent>,
+		username: User['username']
 	): Promise<void> {
+		const foundEvent = await this.calenderEventRepository.findOne({
+			id,
+		})
+
+		if (foundEvent && foundEvent?.username !== username) {
+			throw new UnauthorizedError()
+		}
+
 		await this.calenderEventRepository.update({ ...event, id })
 	}
 
-	public async delete(id: CalenderEvent['id']): Promise<void> {
+	public async delete(
+		id: CalenderEvent['id'],
+		username: User['username']
+	): Promise<void> {
+		const foundEvent = await this.calenderEventRepository.findOne({
+			id,
+		})
+
+		if (foundEvent && foundEvent?.username !== username) {
+			throw new UnauthorizedError()
+		}
+
 		await this.calenderEventRepository.delete(id)
 	}
 }
